refactor: use re-designed account card in account widget
