+++ apps/mobile/src/components/widgets/accounts/components/add-account-card.tsx
@@ -1,0 +1,19 @@
+import { AccountCard } from '@/features/accounts/components/account-card';
+import { t } from '@lingui/macro';
+
+import { PlusIcon } from '@leather.io/ui/native';
+
+export interface AddAccountCardProps {
+  onPress(): void;
+}
+
+export function AddAccountCard({ onPress }: AddAccountCardProps) {
+  return (
+    <AccountCard
+      onPress={onPress}
+      width={200}
+      icon={PlusIcon}
+      primaryTitle={t({ id: 'add_account_card.title', message: 'Add account' })}
+    />
+  );
+}
